+++ MP_Garcia_GeneJoseph_BMIS/Views/AccountView/RegisterAccountView.cs
@@ -68,8 +68,8 @@ namespace MP_Garcia_GeneJoseph_BMIS.Views.AccountView
             this.dsnLlbl = new Label();
 
             // Label for Accounts
-            this.lblResidents.Font = new Font("Trebuchet MS", 14.25F, FontStyle.Regular, GraphicsUnit.Point, ((byte)(0)));
-            this.lblResidents.ForeColor = Color.FromArgb(((int)(((byte)(20)))), ((int)(((byte)(39)))), ((int)(((byte)(78)))));
+            this.lblResidents.Font = new Font("Trebuchet MS", 14.25F, FontStyle.Regular, GraphicsUnit.Point, 0);
+            this.lblResidents.ForeColor = Color.FromArgb(20, 39, 78);
             this.lblResidents.Location = new Point(18, 79);
             this.lblResidents.Size = new Size(179, 24);
             this.lblResidents.AutoSize = true;
@@ -81,33 +81,33 @@ namespace MP_Garcia_GeneJoseph_BMIS.Views.AccountView
             DataGridViewCellStyle dgvcs2 = new DataGridViewCellStyle();
             DataGridViewCellStyle dgvcs3 = new DataGridViewCellStyle();
             DataGridViewCellStyle dgvcs4 = new DataGridViewCellStyle();
-            dgvcs1.BackColor = Color.FromArgb(((int)(((byte)(244)))), ((int)(((byte)(244)))), ((int)(((byte)(244)))));
-            dgvcs1.Font = new Font("Calibri", 11.25F, FontStyle.Regular, GraphicsUnit.Point, ((byte)(0)));
+            dgvcs1.BackColor = Color.FromArgb(244, 244, 244);
+            dgvcs1.Font = new Font("Calibri", 11.25F, FontStyle.Regular, GraphicsUnit.Point, 0);
             dgvcs1.ForeColor = Color.Black;            
             dgvcs2.Alignment = DataGridViewContentAlignment.MiddleCenter;
-            dgvcs2.BackColor = Color.FromArgb(((int)(((byte)(20)))), ((int)(((byte)(39)))), ((int)(((byte)(78)))));
-            dgvcs2.Font = new Font("Calibri", 11.25F, FontStyle.Regular, GraphicsUnit.Point, ((byte)(0)));
-            dgvcs2.ForeColor = Color.FromArgb(((int)(((byte)(241)))), ((int)(((byte)(246)))), ((int)(((byte)(249)))));
-            dgvcs2.SelectionBackColor = Color.FromArgb(((int)(((byte)(217)))), ((int)(((byte)(236)))), ((int)(((byte)(242)))));
+            dgvcs2.BackColor = Color.FromArgb(20, 39, 78);
+            dgvcs2.Font = new Font("Calibri", 11.25F, FontStyle.Regular, GraphicsUnit.Point, 0);
+            dgvcs2.ForeColor = Color.FromArgb(241, 246, 249);
+            dgvcs2.SelectionBackColor = Color.FromArgb(217, 236, 242);
             dgvcs2.SelectionForeColor = SystemColors.Desktop;
             dgvcs2.WrapMode = DataGridViewTriState.True;            
             dgvcs3.Alignment = DataGridViewContentAlignment.MiddleCenter;
             dgvcs3.BackColor = Color.White;
-            dgvcs3.Font = new Font("Calibri", 11.25F, FontStyle.Regular, GraphicsUnit.Point, ((byte)(0)));
+            dgvcs3.Font = new Font("Calibri", 11.25F, FontStyle.Regular, GraphicsUnit.Point, 0);
             dgvcs3.ForeColor = Color.Black;
-            dgvcs3.SelectionBackColor = Color.FromArgb(((int)(((byte)(235)))), ((int)(((byte)(236)))), ((int)(((byte)(241)))));
+            dgvcs3.SelectionBackColor = Color.FromArgb(235, 236, 241);
             dgvcs3.SelectionForeColor = Color.Black;
             dgvcs3.WrapMode = DataGridViewTriState.False;
             dgvcs4.Alignment = DataGridViewContentAlignment.MiddleCenter;
-            dgvcs4.BackColor = Color.FromArgb(((int)(((byte)(155)))), ((int)(((byte)(164)))), ((int)(((byte)(180)))));
-            dgvcs4.Font = new Font("Calibri", 11.25F, FontStyle.Regular, GraphicsUnit.Point, ((byte)(0)));
+            dgvcs4.BackColor = Color.FromArgb(155, 164, 180);
+            dgvcs4.Font = new Font("Calibri", 11.25F, FontStyle.Regular, GraphicsUnit.Point, 0);
             dgvcs4.ForeColor = SystemColors.WindowText;
-            dgvcs4.SelectionBackColor = Color.FromArgb(((int)(((byte)(20)))), ((int)(((byte)(39)))), ((int)(((byte)(78)))));
+            dgvcs4.SelectionBackColor = Color.FromArgb(20, 39, 78);
             dgvcs4.SelectionForeColor = SystemColors.HighlightText;
             dgvcs4.WrapMode = DataGridViewTriState.True;            
             this.dataList.AlternatingRowsDefaultCellStyle = dgvcs1;
             this.dataList.AutoSizeColumnsMode = DataGridViewAutoSizeColumnsMode.Fill;
-            this.dataList.BackgroundColor = Color.FromArgb(((int)(((byte)(241)))), ((int)(((byte)(246)))), ((int)(((byte)(249)))));
+            this.dataList.BackgroundColor = Color.FromArgb(241, 246, 249);
             this.dataList.BorderStyle = BorderStyle.None;
             this.dataList.CellBorderStyle = DataGridViewCellBorderStyle.RaisedHorizontal;
             this.dataList.ClipboardCopyMode = DataGridViewClipboardCopyMode.Disable;
@@ -137,13 +137,13 @@ namespace MP_Garcia_GeneJoseph_BMIS.Views.AccountView
             this.dataList.CellClick += new DataGridViewCellEventHandler(this.RegisterResidentClick);
 
             // Design Components
-            this.dsnBox.BackColor = Color.FromArgb(((int)(((byte)(57)))), ((int)(((byte)(72)))), ((int)(((byte)(103)))));
+            this.dsnBox.BackColor = Color.FromArgb(57, 72, 103);
             this.dsnBox.Location = new Point(-2, 0);
             this.dsnBox.Size = new Size(888, 51);
             this.dsnBox.Name = "dsnBox";
-            this.dsnLlbl.BackColor = Color.FromArgb(((int)(((byte)(57)))), ((int)(((byte)(72)))), ((int)(((byte)(103)))));
-            this.dsnLlbl.Font = new Font("Trebuchet MS", 14.25F, FontStyle.Regular, GraphicsUnit.Point, ((byte)(0)));
-            this.dsnLlbl.ForeColor = Color.FromArgb(((int)(((byte)(241)))), ((int)(((byte)(246)))), ((int)(((byte)(249)))));
+            this.dsnLlbl.BackColor = Color.FromArgb(57, 72, 103);
+            this.dsnLlbl.Font = new Font("Trebuchet MS", 14.25F, FontStyle.Regular, GraphicsUnit.Point, 0);
+            this.dsnLlbl.ForeColor = Color.FromArgb(241, 246, 249);
             this.dsnLlbl.Location = new Point(10, 13);
             this.dsnLlbl.Size = new Size(372, 24);
             this.dsnLlbl.AutoSize = true;
@@ -153,7 +153,7 @@ namespace MP_Garcia_GeneJoseph_BMIS.Views.AccountView
             // Actual Form
             this.AutoScaleDimensions = new SizeF(6F, 13F);
             this.AutoScaleMode = AutoScaleMode.Font;
-            this.BackColor = Color.FromArgb(((int)(((byte)(241)))), ((int)(((byte)(246)))), ((int)(((byte)(249)))));
+            this.BackColor = Color.FromArgb(241, 246, 249);
             this.ClientSize = new Size(884, 561);
             this.FormBorderStyle = FormBorderStyle.FixedToolWindow;
             this.StartPosition = FormStartPosition.CenterScreen;
